Created utils script
